requirements.txt
Django==1.4
South==0.7.5
psycopg2
django-debug-toolbar

-e git://github.com/sunlightlabs/sfapp.git@b8d70e8436ed50620a887845570d1602911b91ec#egg=sfapp-dev

boto==2.5.1
django-storages==1.1.4
jsonfield
django-passwords==0.2.0
django-social-auth
django-registration
python-documentcloud==0.154
django-extensions==0.8
-e hg+https://bitbucket.org/schinckel/django-timedelta-field@0ca08518d3eb60f9a5168b3279dcca1913f6b1f0#egg=django_timedeltafield-0.6.2
-e hg+https://bitbucket.org/schinckel/django-weekday-field@f075f68d54dfe5f4a532b3819c7048a1acc42bc9#egg=django_weekday_field-dev
-e git://github.com/sunlightlabs/django-doccloud.git#egg=django_doccloud-dev
django-reversion==1.6.0
django-ajax-selects==1.2.5
geopy
unicodecsv==0.9.0
name-cleaver==0.3.1
